// apps/hub/src/payload-types.ts
export interface Etiqueta {
  id: number;
  nombre: string;
  /**
   * Nombre alternativo usado en package.json (ej: "react" para React)
   */
  aliasDePaquete?: string | null;
  tipo: 'Tecnologia' | 'Conceptual' | 'Area';
  /**
   * Etiquetas de las que depende esta etiqueta
   */
  dependeDe?: (number | Etiqueta)[] | null;
  /**
   * Directrices que utilizan esta etiqueta
   */
  directricesRelacionadas?: {
    docs?: (number | Directrice)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "directrices".
 */
export interface Directrice {
  id: number;
  /**
   * Título descriptivo de la directriz
   */
  titulo: string;
  /**
   * Contenido completo de la directriz en formato markdown
   */
  contenido: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  /**
   * Indica si la directriz está activa y se incluye en las respuestas
   */
  activa: boolean;
  /**
   * Etiquetas asociadas a esta directriz
   */
  etiquetas: (number | Etiqueta)[];
  updatedAt: string;
  createdAt: string;
}